requirements.txt
asgiref==3.4.1
astroid==2.9.2
colorama==0.4.4
Django==3.2.11
django-ckeditor==6.2.0
django-debug-toolbar==3.2.4
django-js-asset==1.2.2
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
Pillow==9.0.0
platformdirs==2.4.1
pylint==2.12.2
pylint-django==2.5.0
pylint-plugin-utils==0.7
python-memcached==1.59
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typed-ast==1.5.1
typing_extensions==4.0.1
wrapt==1.13.3